Create TopEngagements
Medium/Easy 
Calculate the percentage of search results, out of all the results, that were positioned in the top 3 and clicked by the user.
We have two tables that contain search results. The 'fb_search_results' table contains the search results from a user's search. In this table, search_id is a key that corresponds to the search_id column of the fb_search_events table. The position column refers to the position of the result. The 'fb_search_events' is a table that stores whether or not the user clicked on a particular search result.